Added SimplyTypedLambda
Also abstracted the intsFrom stuff into a package object.

There are a lot of commented items in SimplyTypedLambda.
I left them there because I will try doing the "ExtensibleLambda"
thing and see how that goes; in that case it might be easier
to implement those features (coproducts, ints) on their own
and integrate them into that system.
So it'd be nice to have the code already.

Anyway it makes sense to have the code for when I implement
more complicated type systems that will use them.